# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/{id}": {
      "get": {
        "operationId": "UserController_getUser",
        "summary": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["Users"]
      }
    },
    "/v1/users": {
      "get": {
        "operationId": "UserController_listUsers",
        "summary": "List users",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponseDto"
                }
              }
            }
          }
        },
        "tags": ["Users"]
      }
    }
  },
  "info": {
    "title": "UserTour API",
    "description": "The UserTour API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm9cs634h00001mp50l45n7kz"
          },
          "object": {
            "type": "string",
            "example": "user"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "example": "2022-10-17T12:34:56.000+00:00"
          },
          "companies": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "memberships": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["id", "object", "attributes", "createdAt", "companies", "memberships"]
      },
      "ListUsersResponseDto": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "next": {
            "type": "string",
            "nullable": true,
            "example": "https://api.usertour.com/v1/users/?cursor=cD00ODY%3D"
          },
          "previous": {
            "type": "string",
            "nullable": true,
            "example": "https://api.usertour.com/v1/users/?cursor=cj0xJnA9NDg3"
          }
        },
        "required": ["results", "next", "previous"]
      }
    }
  }
}